use context in mapview, move country polygon switch to its own file
